apps/gmail-email-actions/requirements.txt
# Google APIs
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# HTTP
requests>=2.31.0

# AWS SDK (Lambda runtime provides this, but include for local dev)
boto3>=1.28.0
